+++ src/core/wrapper/layout/middle/middle.center.js
@@ -48,7 +48,7 @@ BI.CenterLayout = BI.inherit(BI.Layout, {
         return {
             type: "bi.grid",
             ref: function (_ref) {
-                self.wrapper = _ref;
+                self.layout = _ref;
             },
             columns: list.length,
             rows: 1,
@@ -65,12 +65,8 @@ BI.CenterLayout = BI.inherit(BI.Layout, {
         throw new Error("不能添加子组件");
     },
 
-    update: function (opt) {
-        return this.wrapper.update(opt);
-    },
-
     populate: function (items) {
-        this.wrapper.populate.apply(this.wrapper, arguments);
+        this.layout.populate.apply(this.layout, arguments);
     }
 });
 BI.shortcut("bi.center", BI.CenterLayout);