+++ src/core/wrapper/layout/middle/middle.float.center.js
@@ -49,7 +49,7 @@ BI.FloatCenterLayout = BI.inherit(BI.Layout, {
         return {
             type: "bi.left",
             ref: function (_ref) {
-                self.wrapper = _ref;
+                self.layout = _ref;
             },
             items: list
         };
@@ -64,12 +64,8 @@ BI.FloatCenterLayout = BI.inherit(BI.Layout, {
         throw new Error("不能添加子组件");
     },
 
-    update: function (opt) {
-        return this.wrapper.update(opt);
-    },
-
     populate: function (items) {
-        this.wrapper.populate.apply(this.wrapper, arguments);
+        this.layout.populate.apply(this.layout, arguments);
     }
 });
 BI.shortcut("bi.float_center", BI.FloatCenterLayout);